+++ AWSCore/Utility/AWSCategory.m
@@ -595,6 +595,12 @@ static NSTimeInterval _clockskew = 0.0;
         || [self isEqualToString:@"ap-east-1"]) {
         return AWSRegionAPEast1;
     }
+    
+    if ([self isEqualToString:@"AWSRegionMESouth1"]
+        || [self isEqualToString:@"MESouth1"]
+        || [self isEqualToString:@"me-south-1"]) {
+        return AWSRegionMESouth1;
+    }
 
     return AWSRegionUnknown;
 }