+++ AWSKinesisVideoArchivedMedia/AWSKinesisVideoArchivedMediaService.m
@@ -26,7 +26,7 @@
 #import "AWSKinesisVideoArchivedMediaResources.h"
 
 static NSString *const AWSInfoKinesisVideoArchivedMedia = @"KinesisVideoArchivedMedia";
-NSString *const AWSKinesisVideoArchivedMediaSDKVersion = @"2.10.0";
+NSString *const AWSKinesisVideoArchivedMediaSDKVersion = @"2.10.1";
 
 
 @interface AWSKinesisVideoArchivedMediaResponseSerializer : AWSJSONResponseSerializer