+++ AWSRekognition/AWSRekognitionService.m
@@ -26,7 +26,7 @@
 #import "AWSRekognitionResources.h"
 
 static NSString *const AWSInfoRekognition = @"Rekognition";
-NSString *const AWSRekognitionSDKVersion = @"2.10.0";
+NSString *const AWSRekognitionSDKVersion = @"2.10.1";
 
 
 @interface AWSRekognitionResponseSerializer : AWSJSONResponseSerializer