+++ AWSS3/AWSS3Model.m
@@ -1086,6 +1086,9 @@ return [date aws_stringValue:AWSDateISO8601DateFormat1];
         if ([value caseInsensitiveCompare:@"ap-east-1"] == NSOrderedSame) {
             return @(AWSS3BucketLocationConstraintAPEast1);
         }
+        if ([value caseInsensitiveCompare:@"me-south-1"] == NSOrderedSame) {
+            return @(AWSS3BucketLocationConstraintMESouth1);
+        }
         return @(AWSS3BucketLocationConstraintUnknown);
     } reverseBlock:^NSString *(NSNumber *value) {
         switch ([value integerValue]) {
@@ -1133,6 +1136,8 @@ return [date aws_stringValue:AWSDateISO8601DateFormat1];
                 return @"eu-north-1";
             case AWSS3BucketLocationConstraintAPEast1:
                 return @"ap-east-1";
+            case AWSS3BucketLocationConstraintMESouth1:
+                return @"me-south-1";
             default:
                 return nil;
         }
@@ -2334,6 +2339,9 @@ return [date aws_stringValue:AWSDateISO8601DateFormat1];
         if ([value caseInsensitiveCompare:@"ap-east-1"] == NSOrderedSame) {
             return @(AWSS3BucketLocationConstraintAPEast1);
         }
+        if ([value caseInsensitiveCompare:@"me-south-1"] == NSOrderedSame) {
+            return @(AWSS3BucketLocationConstraintMESouth1);
+        }
         return @(AWSS3BucketLocationConstraintUnknown);
     } reverseBlock:^NSString *(NSNumber *value) {
         switch ([value integerValue]) {
@@ -2381,6 +2389,8 @@ return [date aws_stringValue:AWSDateISO8601DateFormat1];
                 return @"eu-north-1";
             case AWSS3BucketLocationConstraintAPEast1:
                 return @"ap-east-1";
+            case AWSS3BucketLocationConstraintMESouth1:
+                return @"me-south-1";
             default:
                 return nil;
         }