+++ AWSSES/AWSSESService.m
@@ -26,7 +26,7 @@
 #import "AWSSESResources.h"
 
 static NSString *const AWSInfoSES = @"SES";
-NSString *const AWSSESSDKVersion = @"2.10.0";
+NSString *const AWSSESSDKVersion = @"2.10.1";
 
 
 @interface AWSSESResponseSerializer : AWSXMLResponseSerializer